refactor : 상위 exception 정의
