+++ src/main/java/exception/ClientErrorException.java
@@ -1,20 +1,8 @@
 package exception;
 
-import enums.HttpStatus;
-
-public class ClientErrorException extends RuntimeException {
-    private final ErrorCode errorCode;
+public class ClientErrorException extends ErrorException {
 
     public ClientErrorException(ErrorCode errorCode) {
-        this.errorCode = errorCode;
-    }
-
-    @Override
-    public String getMessage() {
-        return errorCode.getMessage();
-    }
-
-    public HttpStatus getHttpStatus(){
-        return errorCode.getHttpStatus();
+      super(errorCode);
     }
 }
